# openapi.yaml
openapi: 3.1.0
info:
  title: RS3.Dev API
  version: 0.8.0
  summary: Public RuneScape 3 player data API
  description: |
    Draft contract for read-only XP, boss kill-count, collection-log, quest,
    and achievement endpoints.

    Usernames are public lookup keys and are matched case-insensitively. Current
    and previously observed usernames can resolve to the same locally generated
    player UUID.
servers:
  - url: /
    description: Current host
security: []
...
paths:
...
  /v1/reference/skills:
    get:
      tags: [Reference]
      operationId: listSkillReferences
      summary: List skill IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/bosses:
    get:
      tags: [Reference]
      operationId: listBossReferences
      summary: List boss IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/items:
    get:
      tags: [Reference]
      operationId: listItemReferences
      summary: List tracked item IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/boss-log-regions:
    get:
      tags: [Reference]
      operationId: listBossLogRegionReferences
      summary: List boss log region IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/slayer-log-regions:
    get:
      tags: [Reference]
      operationId: listSlayerLogRegionReferences
      summary: List Slayer log region IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/quests:
    get:
      tags: [Reference]
      operationId: listQuestReferences
      summary: List quest IDs and names
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/achievements:
    get:
      tags: [Reference]
      operationId: listAchievementReferences
      summary: List achievement IDs, names, and types
      parameters:
        - $ref: "#/components/parameters/AchievementTypeFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/reference/achievement-types:
    get:
      tags: [Reference]
      operationId: listAchievementTypes
      summary: List achievement types
      responses:
        "200":
          description: Achievement types currently present in the reference data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementTypeCollection"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Number of entries to return per page.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 25
    AchievementTypeFilter:
      name: type
      in: query
      required: false
      description: |
        Normalized achievement type to include. Use `combat` to return only
        Combat Achievements.
      schema:
        type: string
        minLength: 1
        maxLength: 32
      example: combat
...
  responses:
    ReferenceCollection:
      description: A page of numeric ID-to-name reference entries.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReferenceCollection"
    BadRequest:
      description: The request could not be parsed or contains an invalid parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INVALID_REQUEST
            message: One or more request parameters are invalid.
            requestId: 01J67NWYF5W6W5T3XSF8Q7E2QJ
...
    InternalServerError:
      description: An unexpected error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INTERNAL_ERROR
            message: An unexpected error occurred.
            requestId: 01J67NWYF5W6W5T3XSF8Q7E2QJ
  schemas:
...
    Pagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ReferenceCollection:
      type: object
      additionalProperties: false
      required:
        - pagination
        - entries
      properties:
        pagination:
          $ref: "#/components/schemas/Pagination"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceEntry"
    ReferenceEntry:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: integer
          minimum: 0
        name:
          type: string
        type:
          type: string
          minLength: 1
          maxLength: 32
          description: Present for achievement references.
    AchievementTypeCollection:
      type: object
      additionalProperties: false
      required:
        - entries
      properties:
        entries:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - type
            properties:
              type:
                type: string
                minLength: 1
                maxLength: 32
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          description: Identifier used to correlate the response with server logs.